ella/Ciscolan_alakoulun_koodi/requirements.txt
Flask==1.1.2
PyJWT==1.7.1
requests==2.24.0